Moved existing file upload processing from the interfaces layer to the domain layer
