+++ timetable/domain/data_upload_processing/__init__.py
@@ -1,0 +1,2 @@
+"""Convenience imports"""
+from .file_upload_processor import FileUploadProcessor
